+++ lib/cli.ex
@@ -1,6 +1,6 @@
 defmodule Superls.CLI do
   @moduledoc false
-  alias Superls.{CLI.Search, Store, Password, Store}
+  alias Superls.{Store, Password}
 
   use Superls
 
@@ -11,21 +11,35 @@ defmodule Superls.CLI do
   def main(argv) do
     _ = Store.maybe_create_cache_path()
 
-    with {opts, args, []} <-
-           OptionParser.parse(argv, aliases: [p: :password], strict: @options),
-         is_password? <- Keyword.get(opts, :password, false),
-         password <- if(is_password?, do: Password.io_get(), else: "") do
-      main_args(args, password)
-    else
-      {_, _, wrong} ->
-        IO.puts("invalid command, check #{inspect(wrong)}")
-    end
-  catch
-    err -> IO.puts("#{err}")
+    {opts, args, []} = OptionParser.parse(argv, aliases: [p: :password], strict: @options)
+    main_args(args, opts)
   end
 
-  defp main_args(["index", vol_path, store], password) do
-    Store.archive(vol_path, store_name(store), password)
+  defp main_args(["index" | args], opts) do
+    {vol_path, store_name} =
+      case args do
+        [vol_path, store] ->
+          {vol_path, store}
+
+        [vol_path] ->
+          {vol_path, default_store()}
+
+        _ ->
+          IO.puts("** missing volume path\n\ntry: superls help")
+          throw({:error, :volume_path_notfound})
+      end
+
+    if !File.dir?(vol_path) do
+      IO.puts("** error: #{vol_path} is an invalid directory\n\ntry: superls help")
+      throw({:error, :invalid_directory})
+    end
+
+    is_password? = Keyword.get(opts, :password, false)
+    password = if(is_password?, do: Password.io_get(), else: "")
+
+    Store.archive(vol_path, store_name, password)
+  catch
+    error -> error
   end
 
   defp main_args(["help"], _passwd) do
@@ -49,39 +63,60 @@ defmodule Superls.CLI do
   end
 
   # search in store
-  defp main_args([store], password) do
-    store_name = store_name(store)
-    mi = Store.get_merged_index_from_store(store_name, password)
-    :ok = :shell.start_interactive({Search, :start, [mi, [store: store_name, passwd: password]]})
+  defp main_args(args, passwd) do
+    passwd = normalize_passwd(passwd)
+    store_name = valid_name(args)
 
-    :timer.sleep(:infinity)
-  rescue
-    _e in File.Error ->
-      default_store = default_store()
+    try do
+      mi = Store.get_merged_index_from_store(store_name, passwd)
+      :ok = start_interactive([mi, [store: store_name, passwd: passwd]])
+    rescue
+      _e in File.Error ->
+        default_store = default_store()
 
-      store =
-        case store_name(store) do
-          ^default_store -> ""
-          any -> any
-        end
+        store =
+          case store_name do
+            ^default_store -> ""
+            any -> any
+          end
 
-      IO.puts("""
-      ** index `#{store_name(store)}` is missing.
-      first create index with: superls index /path/to/my/volume_files -i #{store} [- p]
-      """)
-  catch
-    :enoent ->
-      IO.puts("** index `#{store_name(store)}` not found\n\ntry: superls help")
+        IO.puts("""
+        ** index `#{store_name}` is missing.
+        first create index with: superls index /path/to/my/volume_files -i #{store} [- p]
+        """)
 
-    err ->
-      IO.puts("** #{err}")
-      main_args([store], Password.io_get())
+        {:error, :store_missing}
+    catch
+      :enoent ->
+        IO.puts("** index `#{store_name}` not found\n\ntry: superls help")
+        {:error, :enoent}
+
+      err ->
+        IO.puts("#{err}")
+        retry_with_password(args)
+    end
   end
 
-  defp main_args(args, _password) do
-    IO.puts("** unknown command:\n  #{Enum.join(["superls" | args], " ")}\ntype `superls help`")
+  if Mix.env() == :test do
+    defp retry_with_password(args) do
+      main_args(args, "secret")
+    end
+
+    defp start_interactive(_args), do: :ok
+  else
+    defp retry_with_password(args) do
+      main_args(args, Password.io_get())
+    end
+
+    defp start_interactive(args) do
+      :ok = :shell.start_interactive({Superls.CLI.Search, :start, args})
+      :timer.sleep(:infinity)
+    end
   end
 
-  defp store_name(""), do: default_store()
-  defp store_name(str) when is_binary(str), do: str
+  defp normalize_passwd(passwd) when is_list(passwd), do: ""
+  defp normalize_passwd(passwd), do: passwd
+
+  defp valid_name([]), do: default_store()
+  defp valid_name([str]), do: str
 end