+++ CHANGELOG.md
@@ -5,8 +5,7 @@
  * use native compression for stored indexes (require re-archiving)
  * fix wrong password tests
  * properly decode stored index
- * rename the archive command by index
- * rename flag -s by -i to specify an index
+ * rework the command line
 
 ## v1.3.0  (2025-09-25)
  * rely on otp-28 interactive shell with autocompletion and history